refactor!: Relocate CLI modules below to `aws_annoying.cli` package
