+++ aws_annoying/cli/main.py
@@ -1,10 +1,10 @@
 # flake8: noqa: F401
 from __future__ import annotations
 
-import aws_annoying.ecs_task_definition_lifecycle
-import aws_annoying.load_variables
-import aws_annoying.mfa
-import aws_annoying.session_manager
+import aws_annoying.cli.ecs_task_definition_lifecycle
+import aws_annoying.cli.load_variables
+import aws_annoying.cli.mfa
+import aws_annoying.cli.session_manager
 from aws_annoying.utils.debugger import input_as_args
 
 # App with all commands registered
