+++ aws_annoying/cli/session_manager/__init__.py
