+++ aws_annoying/cli/session_manager/_app.py
@@ -1,6 +1,6 @@
 import typer
 
-from aws_annoying.app import app
+from aws_annoying.cli.app import app
 
 session_manager_app = typer.Typer(
     no_args_is_help=True,
