+++ aws_annoying/cli/session_manager/errors.py
