+++ aws_annoying/cli/session_manager/install.py
