+++ aws_annoying/cli/session_manager/session_manager.py
