+++ aws_annoying/cli/session_manager/stop.py
