+++ tests/cli/_helpers/__init__.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 
 from .command_builder import repeat_options
+from .invoke import invoke_cli
 from .string_ import normalize_console_output
 
-__all__ = ("PRINTENV_PY", "normalize_console_output", "printenv_py", "repeat_options")
+__all__ = ("PRINTENV_PY", "invoke_cli", "normalize_console_output", "printenv_py", "repeat_options")
 
 PRINTENV_PY = (Path(__file__).parent / "scripts" / "printenv.py").absolute()
