+++ tests/cli/_helpers/command_builder.py
