+++ tests/cli/_helpers/scripts/printenv.py
