+++ tests/cli/_helpers/string_.py
