+++ tests/cli/mfa/snapshots/test_configure/test_basic/persist/aws_config.ini
