+++ tests/cli/mfa/snapshots/test_configure/test_basic/skip_persist/stdout.txt
