+++ tests/cli/mfa/snapshots/test_configure/test_load_existing_config/stdout.txt
