+++ tests/cli/session_manager/test_session_manager.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import pytest
 from typer.testing import CliRunner
 
-from aws_annoying.main import app
-from aws_annoying.session_manager.session_manager import SessionManager
+from aws_annoying.cli.main import app
+from aws_annoying.cli.session_manager.session_manager import SessionManager
 from aws_annoying.utils.downloader import DummyDownloader
 
 runner = CliRunner()
