+++ tests/cli/snapshots/test_ecs_task_definition_lifecycle/test_basic/stdout.txt
