+++ tests/cli/snapshots/test_ecs_task_definition_lifecycle/test_dry_run/stdout.txt
