+++ tests/cli/snapshots/test_load_variables/test_nothing/stdout.txt
