+++ tests/cli/snapshots/test_load_variables/test_overwrite_env/stdout.txt
@@ -9,7 +9,7 @@
 └──────────────┴───────────────────────────────────────────────────────────────┘
 🔍 Retrieving variables from AWS resources...
 ✅ Retrieved 1 secrets and 2 parameters.
-🚀 Running the command: tests/_helpers/scripts/printenv.py DJANGO_SETTINGS_MODULE DJANGO_SECRET_KEY DJANGO_DEBUG DJANGO_ALLOWED_HOSTS
+🚀 Running the command: tests/cli/_helpers/scripts/printenv.py DJANGO_SETTINGS_MODULE DJANGO_SECRET_KEY DJANGO_DEBUG DJANGO_ALLOWED_HOSTS
 DJANGO_SETTINGS_MODULE=config.settings.local
 DJANGO_SECRET_KEY=my-secret-key
 DJANGO_DEBUG=False
