+++ tests/cli/snapshots/test_load_variables/test_replace_quiet/stdout.txt
