+++ tests/cli/snapshots/test_load_variables/test_resource_not_found/ssm/stdout.txt
