+++ tests/cli/snapshots/test_load_variables/test_unsupported_resource/stdout.txt
